+++ src/config/socket.rs
@@ -1,6 +1,7 @@
 use serde::Deserialize;
 
 #[derive(Debug, Deserialize, Clone)]
+#[allow(dead_code)]
 pub struct SocketConfig {
     pub path: String,
     pub group: String,